requirements.txt
streamlit
transformers
torch
nltk
spacy
google-generativeai
python-dotenv
scikit-learn
